requirements.txt
mss
opencv-python
pyautogui
pynput
numpy
rich
pillow
tk
pytesseract
uiautomation

# Optional helpers used by image-based detection
scipy
opencv-contrib-python